requirements.txt
absl-py
torch-scatter~=2.1.0+pt113cu117
torch-sparse~=0.6.16+pt113cu117
torch-geometric
torch~=1.13.1+cu117
numpy~=1.24.1